reuse createAttributesObject in r-avatar
