+++ addon/mixins/r-button-mixin.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
-import { getAttributeValue } from 'ui-library/utils/get-attribute-value';
+import { createAttributesObject } from 'ui-library/utils/create-attributes-object';
 
-const { Mixin, get, setProperties, getProperties } = Ember;
+const { Mixin, get, setProperties } = Ember;
 
 export default Mixin.create({
   classNames: ['r-btn'],
@@ -44,34 +44,7 @@ export default Mixin.create({
     const defaults = get(this, 'defaults');
     const suppliedAttrs = get(this, 'attrs');
 
-    const attrObj = this._createAttributesObject(allowedAttributes, defaults, suppliedAttrs);
+    const attrObj = createAttributesObject(allowedAttributes, defaults, suppliedAttrs);
     setProperties(this, attrObj);
-  },
-
-  /**
-  * @desc creates an object used to safely set attributes
-  * @param {Object} allowedAttributes - valid attribute values
-  * @param {Object} defaults - fallback attributes if valid attrs not provided
-  * @param {Object} suppliedAttrs - the attributes passed to the component
-  * @returns {Object} Object of attribute keys with true value
-  */
-  _createAttributesObject(allowedAttributes, defaults, suppliedAttrs = {}) {
-    const dataObj = {};
-
-    //Filter suppliedAttrs to just valid ones. E.g. 'label' is not valid
-    const validSuppliedAttrKeys = Object.keys(suppliedAttrs)
-      .filter(attr => Object.keys(allowedAttributes).indexOf(attr) > -1);
-
-    const validSuppliedAttrs = getProperties(suppliedAttrs, validSuppliedAttrKeys);
-
-    Object.keys(allowedAttributes).map(attribute => {
-      const property = getAttributeValue(
-        allowedAttributes[attribute],
-        defaults[attribute],
-        get(validSuppliedAttrs, attribute)
-      );
-      dataObj[property] = true;
-    });
-    return dataObj;
   }
 });
